fix AOP AspectJ, due to using forceAjcCompile has problem with Spring AOP
